refactor: extract modular viz package from map feature
Restructure map visualizations into a self-contained viz/ package
that can be extracted as @quodeq/viz library.

Structure:
  viz/
    index.js              — barrel export for all components and utilities
    components/           — GalaxyView, GalaxyFolderView, ZoomablePackView,
                            RiskMatrixView, HeatGridView, FileShape, HeatGridCells
    core/                 — galaxyCore.js, mapColors.js, fileTree.js

Changes:
- Deduplicate GalaxyView: removed ~140 lines of copied code, now imports
  from galaxyCore.js (parseCSSColor, getThemeColors, scoreRGB, sevRGB,
  rgb, rgba, drawGlow, drawParticles, mkParticles, LEGEND_ITEMS)
- Extract API call: GalaxyView no longer imports from api/standards.js,
  accepts standardTypes as prop instead. MapPage handles the fetch.
- Add gradeToScore export to galaxyCore.js
- Co-locate HeatGridCells in viz/components/ for self-containment
- All viz components have clean prop interfaces with no app-level imports
- MapPage is now a thin orchestrator wrapper
