+++ src/quodeq/ui/src/features/map/viz/components/FileShape.jsx
