+++ src/quodeq/ui/src/features/map/viz/components/GalaxyView.jsx
@@ -1,74 +1,9 @@
 import { useRef, useEffect, useMemo, useCallback, useState } from 'react';
-import { seedHash, seededRng } from './galaxyCore.js';
-
-const TAU = Math.PI * 2;
-
-function parseCSSColor(cssColor) {
-  // Parse any CSS color string to {r,g,b} via a temporary canvas
-  if (!parseCSSColor._ctx) {
-    const c = document.createElement('canvas');
-    c.width = c.height = 1;
-    parseCSSColor._ctx = c.getContext('2d');
-  }
-  const ctx = parseCSSColor._ctx;
-  ctx.clearRect(0, 0, 1, 1);
-  ctx.fillStyle = cssColor;
-  ctx.fillRect(0, 0, 1, 1);
-  const [r, g, b] = ctx.getImageData(0, 0, 1, 1).data;
-  return { r, g, b };
-}
-
-let _themeColors = null;
-function getThemeColors() {
-  if (_themeColors) return _themeColors;
-  const style = getComputedStyle(document.documentElement);
-  const get = (v) => parseCSSColor(style.getPropertyValue(v).trim() || '#888');
-  const raw = (v) => style.getPropertyValue(v).trim();
-  _themeColors = {
-    critical: get('--color-sev-critical-text'),
-    major: get('--color-sev-major-text'),
-    minor: get('--color-sev-minor-text'),
-    compliance: get('--color-compliance'),
-    gradeTop: get('--color-grade-top-text'),
-    gradeHigh: get('--color-grade-high-text'),
-    gradeMid: get('--color-grade-mid-text'),
-    gradeLow: get('--color-grade-low-text'),
-    gradeBottom: get('--color-grade-bottom-text'),
-    bg: raw('--color-bg') || '#0a0a16',
-    bgAlt: raw('--color-surface-alt') || '#12122a',
-    text: get('--color-text'),
-    textMuted: get('--color-text-muted'),
-    border: get('--color-border'),
-    surface: raw('--color-surface') || '#1a1a2e',
-  };
-  // Invalidate on theme change (class changes on root)
-  const obs = new MutationObserver(() => { _themeColors = null; });
-  obs.observe(document.documentElement, { attributes: true, attributeFilter: ['class', 'data-theme'] });
-  return _themeColors;
-}
-
-// score is 0-10 scale (matching the app's grading system)
-function scoreRGB(score) {
-  const tc = getThemeColors();
-  if (score >= 9) return tc.gradeTop;
-  if (score >= 7) return tc.gradeHigh;
-  if (score >= 5) return tc.gradeMid;
-  if (score >= 3) return tc.gradeLow;
-  return tc.gradeBottom;
-}
-
-// Returns 0-10 scale to match scoreRGB thresholds
-function gradeToScore(grade) {
-  return { A: 9.5, B: 8, C: 6.5, D: 4.5, F: 2 }[grade] || 5;
-}
-
-function sevRGB(sev) {
-  const tc = getThemeColors();
-  return tc[sev] || tc.minor;
-}
-
-function rgb(c) { return `rgb(${c.r},${c.g},${c.b})`; }
-function rgba(c, a) { return `rgba(${c.r},${c.g},${c.b},${a})`; }
+import {
+  TAU, getThemeColors, scoreRGB, sevRGB, rgb, rgba,
+  drawGlow, drawParticles, mkParticles,
+  seedHash, seededRng, gradeToScore, LEGEND_ITEMS,
+} from '../core/galaxyCore.js';
 
 /** Group violations and compliance by principle name, returning { [principleName]: { violations, compliance } } */
 function groupByPrinciple(dim) {
@@ -113,28 +48,6 @@ function computePrincipleScore(rawScore, grade, violationCount, complianceCount)
   if (grade) return gradeToScore(grade);
   const total = violationCount + complianceCount;
   return total > 0 ? (complianceCount / total) * 10 : 5;
-}
-
-function mkParticles(critical, major, minor, baseRadius) {
-  const ps = [];
-  const add = (n, sev) => {
-    const col = sevRGB(sev);
-    for (let i = 0; i < Math.min(n, 10); i++) {
-      ps.push({
-        col, sev,
-        or: baseRadius + 22 + Math.random() * 28,
-        os: (0.03 + Math.random() * 0.07) * (Math.random() > 0.5 ? 1 : -1),
-        op: Math.random() * TAU,
-        sz: sev === 'critical' ? 3.2 + Math.random() * 0.8 : sev === 'major' ? 2.6 + Math.random() * 0.5 : 1.8 + Math.random() * 0.5,
-        ec: 0.65 + Math.random() * 0.35,
-        tp: Math.random() * TAU,
-      });
-    }
-  };
-  add(critical, 'critical');
-  add(major, 'major');
-  add(minor, 'minor');
-  return ps;
 }
 
 const CONSTELLATION_LABELS = {
@@ -389,33 +302,13 @@ function buildScene(dimensions, W, H, standardTypes) {
   return { stars, principles, connections, constellations, bg, _maxExtent };
 }
 
-const LEGEND_ITEMS = [
-  { color: 'var(--color-grade-top-text)', label: 'Exemplary' },
-  { color: 'var(--color-grade-high-text)', label: 'Good' },
-  { color: 'var(--color-grade-mid-text)', label: 'Adequate' },
-  { color: 'var(--color-grade-low-text)', label: 'Poor' },
-  { color: 'var(--color-grade-bottom-text)', label: 'Critical' },
-];
-
 // Module-level saved state — persists across unmount/remount (back from detail)
 let _savedGalaxyNav = null;
 let _savedGalaxyCam = null;
 
-export default function GalaxyView({ dimensions, onNavigate, showLabels = true, setShowLabels, resetKey = 0, projectName = '' }) {
+export default function GalaxyView({ dimensions, onNavigate, showLabels = true, setShowLabels, resetKey = 0, projectName = '', standardTypes = {} }) {
   const canvasRef = useRef(null);
   const [size, setSize] = useState({ w: 800, h: 600 });
-
-  // Fetch standard types for constellation grouping
-  const [standardTypes, setStandardTypes] = useState({});
-  useEffect(() => {
-    import('../../../api/standards.js').then(({ listStandards }) => {
-      listStandards().then(stds => {
-        const map = {};
-        stds.forEach(s => { map[(s.id || '').toLowerCase()] = s.type || 'custom'; });
-        setStandardTypes(map);
-      });
-    }).catch(() => {});
-  }, []);
 
   // Build layout once when dimension structure or standard types change
   const dimKey = useMemo(() => dimensions.map(d => d.dimension).sort().join('|'), [dimensions]);
@@ -560,38 +453,6 @@ export default function GalaxyView({ dimensions, onNavigate, showLabels = true, 
     if (nav.depth === 2 && nav.dim !== null && nav.prin !== null) { const p = scene.principles[nav.dim][nav.prin]; return { x: p.x, y: p.y, z: 50 }; }
     return camRef.current;
   }, [scene, size.w, size.h, getFitZoom]);
-
-  // Drawing helpers
-  const drawGlow = useCallback((ctx, sx, sy, sr, col, alpha) => {
-    if (sr < 0.3 || alpha < 0.01) return;
-    const { r, g, b } = col;
-    const gl = ctx.createRadialGradient(sx, sy, sr * 0.4, sx, sy, sr * 3);
-    gl.addColorStop(0, `rgba(${r},${g},${b},${0.15 * alpha})`);
-    gl.addColorStop(1, `rgba(${r},${g},${b},0)`);
-    ctx.beginPath(); ctx.arc(sx, sy, sr * 3, 0, TAU); ctx.fillStyle = gl; ctx.fill();
-    const co = ctx.createRadialGradient(sx, sy, 0, sx, sy, sr);
-    co.addColorStop(0, `rgba(${Math.min(255, r + 60)},${Math.min(255, g + 60)},${Math.min(255, b + 60)},${0.9 * alpha})`);
-    co.addColorStop(0.6, `rgba(${r},${g},${b},${0.6 * alpha})`);
-    co.addColorStop(1, `rgba(${r},${g},${b},${0.15 * alpha})`);
-    ctx.beginPath(); ctx.arc(sx, sy, sr, 0, TAU); ctx.fillStyle = co; ctx.fill();
-  }, []);
-
-  const drawParticles = useCallback((ctx, cx, cy, particles, scale, alpha, t, drawScale) => {
-    const ds = drawScale ?? scale;
-    particles.forEach(p => {
-      const a = t * p.os + p.op;
-      const px = cx + Math.cos(a) * p.or * p.ec * scale;
-      const py = cy + Math.sin(a) * p.or * scale;
-      const tw = 0.5 + 0.08 * Math.sin(t * 0.5 + p.tp);
-      const sz = p.sz * ds;
-      if (sz < 0.15) return;
-      const { r, g, b } = p.col;
-      ctx.beginPath(); ctx.arc(px, py, sz * 3, 0, TAU);
-      ctx.fillStyle = `rgba(${r},${g},${b},${tw * 0.12 * alpha})`; ctx.fill();
-      ctx.beginPath(); ctx.arc(px, py, sz, 0, TAU);
-      ctx.fillStyle = `rgba(${r},${g},${b},${(tw + 0.2) * alpha})`; ctx.fill();
-    });
-  }, []);
 
   // --- Main animation loop ---
   useEffect(() => {
@@ -843,7 +704,7 @@ export default function GalaxyView({ dimensions, onNavigate, showLabels = true, 
 
     frameRef.current = requestAnimationFrame(frame);
     return () => { running = false; cancelAnimationFrame(frameRef.current); };
-  }, [scene, size, showLabels, w2s, getTarget, drawGlow, drawParticles]);
+  }, [scene, size, showLabels, w2s, getTarget]);
 
   // --- Mouse + click handlers ---
   const handleMouseMove = useCallback((e) => {
