+++ src/quodeq/ui/src/features/map/viz/components/HeatGridCells.jsx
@@ -1,4 +1,4 @@
-import { severityCellStyle, complianceRateCellStyle } from '../features/map/viz/core/mapColors.js';
+import { severityCellStyle, complianceRateCellStyle } from '../core/mapColors.js';
 
 const SEVERITY_LEVELS = ['critical', 'major', 'minor'];
