+++ src/quodeq/ui/src/features/map/viz/core/mapColors.js
