+++ src/quodeq/ui/src/features/map/viz/index.js
@@ -1,0 +1,9 @@
+export { default as GalaxyView } from './components/GalaxyView.jsx';
+export { default as GalaxyFolderView } from './components/GalaxyFolderView.jsx';
+export { default as ZoomablePackView, resetSavedFocus } from './components/ZoomablePackView.jsx';
+export { default as RiskMatrixView } from './components/RiskMatrixView.jsx';
+export { default as HeatGridView } from './components/HeatGridView.jsx';
+export { default as FileShape } from './components/FileShape.jsx';
+export { buildFileTree, treeNodeToFileObj } from './core/fileTree.js';
+export * from './core/mapColors.js';
+export * from './core/galaxyCore.js';
